Minor unicode handling fixes

+++ src/xrCore/XML/XMLDocument.cpp
@@ -22,7 +22,7 @@ enum class ParseIncludeResult
 ParseIncludeResult ParseInclude(pstr string, pcstr& out_include_name)
 {
     // Skip any whitespace characters
-    while (*string != '\0' && std::isblank(*string))
+    while (*string != '\0' && std::isblank(static_cast<u8> (*string)))
     {
         ++string;
     }
@@ -35,7 +35,7 @@ ParseIncludeResult ParseInclude(pstr string, pcstr& out_include_name)
     string += 8;
 
     // Skip any whitespace characters
-    while (*string != '\0' && std::isblank(*string))
+    while (*string != '\0' && std::isblank(static_cast<u8>(*string)))
         ++string;
 
     // Check that after the tag there is a quote
